+++ NEWS.md
@@ -1,3 +1,11 @@
+# malariasimulation 1.3.0
+
+  * Custom demography
+  * Rainfall floor
+  * Documentation fixes for vector controls
+  * Simple model for severe disease
+  * Alignment of immunity function with legacy model
+
 # malariasimulation 1.2.0
 
   * added a `news.md` file to track changes to the package.